pyproject.toml
[tool.poetry]
name = "practic-tg-bot"
version = "0.1.0"
description = ""
authors = ["Michael <pozdinmisha666@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
